memory_console/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
google-cloud-firestore==2.21.0
requests==2.32.5
